requirements.txt
pillow
python-dotenv
keyboard
requests
colorama
SpeechRecognition
google.generativeai
websocket-client
pyaudio